Add files via upload
The main aim of Agriculture Crop Order Management system is to help farmers in the field of agriculture. The main objective of this system is to avoid the brokerage system by providing direct marketing facility to sell crops and also provide information to customer about different variety of crops available in the market.